Created the class which implements the behavior of the interface methods and solved the TODO's
Please note that the getter methods were added to keep the program functioning properly, they will be addressed in a subsequent pull-request (this is the reason why I didn't add any comments on them). Furthermore, the variables are duplicated from the StartUp class for the ease of construction, these too will be addressed where needed. Also, new pull request will be opened with the new similar interfaces.